+++ test/design_mpc_implementation_test.jl
@@ -1065,7 +1065,7 @@ end
     @test C_densenet_nl.system == QTP_sys_densenet
     @test typeof(C_densenet_nl.tuning.modeler) == JuMP.Model
     @test JuMP.solver_name(C_densenet_nl.tuning.modeler) == "Ipopt"
-    @test length(JuMP.object_dictionary(C_densenet_nl.tuning.modeler)) == 8
+    @test length(JuMP.object_dictionary(C_densenet_nl.tuning.modeler)) == 10
     @test size(JuMP.object_dictionary(C_densenet_nl.tuning.modeler)[:u_reference]) ==
           (QTP_sys_densenet.inputdim, horizon)
     @test size(JuMP.object_dictionary(C_densenet_nl.tuning.modeler)[:u]) ==
@@ -1162,7 +1162,7 @@ end
     @test C_densenet_milp.system == QTP_sys_densenet
     @test typeof(C_densenet_milp.tuning.modeler) == JuMP.Model
     @test JuMP.solver_name(C_densenet_milp.tuning.modeler) == "SCIP"
-    @test length(JuMP.object_dictionary(C_densenet_milp.tuning.modeler)) == 11
+    @test length(JuMP.object_dictionary(C_densenet_milp.tuning.modeler)) == 13
     @test size(JuMP.object_dictionary(C_densenet_milp.tuning.modeler)[:u_reference]) ==
           (QTP_sys_densenet.inputdim, horizon)
     @test size(JuMP.object_dictionary(C_densenet_milp.tuning.modeler)[:u]) ==
@@ -1809,7 +1809,7 @@ end
 
 end
 
-@testset "design linear Model Predictive Control with neuralnetODE_type1 model" begin
+@testset "design linear Model Predictive Control with NeuralODE model" begin
 
     hmin = 0.2
     h1max = 1.36
@@ -1829,19 +1829,19 @@ end
 
     #get the resnet model to design the mpc controler
     neuralnetODE_type1_machine =
-        machine("./models_saved/neuralnetODE_type1_train_result.jls")
+        machine("./models_saved/NeuralODE_train_result.jls")
 
     #extract best model from the all trained models
     mlj_neuralnetODE_type1 =
         fitted_params(fitted_params(neuralnetODE_type1_machine).machine).best_model
-    f_neuralnetODE_type1 =
+    f_neuralODE =
         fitted_params(fitted_params(neuralnetODE_type1_machine).machine).best_fitted_params[1]
-    type_neuralnetODE_type1 = mlj_neuralnetODE_type1.builder
+    type_neuralODE = mlj_neuralnetODE_type1.builder
 
     #system definition with MAthematical systems
     QTP_sys_neuralnetODE_type1 =
         MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem(
-            f_neuralnetODE_type1,
+            f_neuralODE,
             4,
             2,
             x_cons,
@@ -1919,4 +1919,333 @@ end
 
 end
 
+@testset "design linear Model Predictive Control with Rknn1 model" begin
+
+      hmin = 0.2
+      h1max = 1.36
+      h2max = 1.36
+      h3max = 1.30
+      h4max = 1.30
+      qmin = 0
+      qamax = 4
+      qbmax = 3.26
+  
+      #Constraint definition:
+      x_cons = LazySets.Hyperrectangle(
+          low = [hmin, hmin, hmin, hmin],
+          high = [h1max, h2max, h3max, h4max],
+      )
+      u_cons = LazySets.Hyperrectangle(low = [qmin, qmin], high = [qamax, qbmax])
+  
+      #get the resnet model to design the mpc controler
+      rknn1_machine =
+          machine("./models_saved/rknn1_train_result.jls")
+  
+      #extract best model from the all trained models
+      mlj_rknn1 =
+          fitted_params(fitted_params(rknn1_machine).machine).best_model
+      f_rknn1 =
+          fitted_params(fitted_params(rknn1_machine).machine).best_fitted_params[1]
+      type_rknn1 = mlj_rknn1.builder
+  
+      #system definition with MAthematical systems
+      QTP_sys_rknn1 =
+          MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem(
+              f_rknn1,
+              4,
+              2,
+              x_cons,
+              u_cons,
+          )
+  
+      #MPC design parameters
+      horizon = 15
+      sample_time = 5
+      terminal_ingredient = false
+      max_time = 5
+      x = [0.65, 0.65, 0.65, 0.65] * ones(1, horizon + 1)
+      u = [1.2, 1.2] * ones(1, horizon)
+      references = AutomationLabsModelPredictiveControl.ReferencesStateInput(x, u)
+  
+      ##############################
+      ###           neuralnetODE_type1          ###                    
+      ##############################
+  
+      ### neuralnetODE_type1 L MPC ###
+      method = AutomationLabsModelPredictiveControl.LinearProgramming()
+  
+      C_rknn1_linear = _model_predictive_control_design(
+          QTP_sys_rknn1,
+          horizon,
+          sample_time,
+          references;
+          mpc_programming_type = "linear",
+      )
+  
+      ### start evaluate neuralnetODE_type1 L MPC implementation ###
+      @test C_rknn1_linear.system == QTP_sys_rknn1
+      @test typeof(C_rknn1_linear.tuning.modeler) == JuMP.Model
+      @test JuMP.solver_name(C_rknn1_linear.tuning.modeler) == "SCIP"
+      @test length(JuMP.object_dictionary(C_rknn1_linear.tuning.modeler)) == 6
+      @test size(
+          JuMP.object_dictionary(C_rknn1_linear.tuning.modeler)[:u_reference],
+      ) == (QTP_sys_rknn1.inputdim, horizon)
+      @test size(JuMP.object_dictionary(C_rknn1_linear.tuning.modeler)[:u]) ==
+            (QTP_sys_rknn1.inputdim, horizon)
+      @test size(JuMP.object_dictionary(C_rknn1_linear.tuning.modeler)[:e_u]) ==
+            (QTP_sys_rknn1.inputdim, horizon)
+      @test size(
+          JuMP.object_dictionary(C_rknn1_linear.tuning.modeler)[:x_reference],
+      ) == (QTP_sys_rknn1.statedim, horizon + 1)
+      @test size(JuMP.object_dictionary(C_rknn1_linear.tuning.modeler)[:x]) ==
+            (QTP_sys_rknn1.statedim, horizon + 1)
+      @test size(JuMP.object_dictionary(C_rknn1_linear.tuning.modeler)[:e_x]) ==
+            (QTP_sys_rknn1.statedim, horizon + 1)
+      @test JuMP.objective_function(C_rknn1_linear.tuning.modeler) != 0 #to improve
+      @test JuMP.objective_function_type(C_rknn1_linear.tuning.modeler) ==
+            JuMP.QuadExpr
+      @test JuMP.list_of_constraint_types(C_rknn1_linear.tuning.modeler) == [
+          (AffExpr, MathOptInterface.EqualTo{Float64}),
+          (AffExpr, MathOptInterface.LessThan{Float64}),
+          (VariableRef, MathOptInterface.EqualTo{Float64}),
+      ]
+      @test C_rknn1_linear.tuning.reference == references
+      @test C_rknn1_linear.tuning.horizon == horizon
+      @test C_rknn1_linear.tuning.terminal_ingredient.Xf == "none"
+      @test C_rknn1_linear.tuning.terminal_ingredient.P != 0
+      @test C_rknn1_linear.tuning.sample_time == 5.0
+      @test C_rknn1_linear.tuning.max_time == 30 #to modify
+      @test size(C_rknn1_linear.initialization) ==
+            (QTP_sys_rknn1.statedim,)
+      @test size(C_rknn1_linear.computation_results.x) ==
+            (QTP_sys_rknn1.statedim, horizon + 1)
+      @test size(C_rknn1_linear.computation_results.e_x) ==
+            (QTP_sys_rknn1.statedim, horizon + 1)
+      @test size(C_rknn1_linear.computation_results.u) ==
+            (QTP_sys_rknn1.inputdim, horizon)
+      @test size(C_rknn1_linear.computation_results.e_u) ==
+            (QTP_sys_rknn1.inputdim, horizon)
+      ### end evaluate neuralnetODE_type1 L MPC implementation ###
+  
+  end
+
+  @testset "design linear Model Predictive Control with rknn2 model" begin
+
+      hmin = 0.2
+      h1max = 1.36
+      h2max = 1.36
+      h3max = 1.30
+      h4max = 1.30
+      qmin = 0
+      qamax = 4
+      qbmax = 3.26
+  
+      #Constraint definition:
+      x_cons = LazySets.Hyperrectangle(
+          low = [hmin, hmin, hmin, hmin],
+          high = [h1max, h2max, h3max, h4max],
+      )
+      u_cons = LazySets.Hyperrectangle(low = [qmin, qmin], high = [qamax, qbmax])
+  
+      #get the resnet model to design the mpc controler
+      rknn2_machine =
+          machine("./models_saved/rknn2_train_result.jls")
+  
+      #extract best model from the all trained models
+      mlj_rknn2 =
+          fitted_params(fitted_params(rknn2_machine).machine).best_model
+      f_rknn2 =
+          fitted_params(fitted_params(rknn2_machine).machine).best_fitted_params[1]
+      type_rknn2 = mlj_rknn2.builder
+  
+      #system definition with MAthematical systems
+      QTP_sys_rknn2 =
+          MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem(
+              f_rknn2,
+              4,
+              2,
+              x_cons,
+              u_cons,
+          )
+  
+      #MPC design parameters
+      horizon = 15
+      sample_time = 5
+      terminal_ingredient = false
+      max_time = 5
+      x = [0.65, 0.65, 0.65, 0.65] * ones(1, horizon + 1)
+      u = [1.2, 1.2] * ones(1, horizon)
+      references = AutomationLabsModelPredictiveControl.ReferencesStateInput(x, u)
+  
+      ##############################
+      ###           neuralnetODE_type1          ###                    
+      ##############################
+  
+      ### neuralnetODE_type1 L MPC ###
+      method = AutomationLabsModelPredictiveControl.LinearProgramming()
+  
+      C_rknn2_linear = _model_predictive_control_design(
+          QTP_sys_rknn2,
+          horizon,
+          sample_time,
+          references;
+          mpc_programming_type = "linear",
+      )
+  
+      ### start evaluate neuralnetODE_type1 L MPC implementation ###
+      @test C_rknn2_linear.system == QTP_sys_rknn2
+      @test typeof(C_rknn2_linear.tuning.modeler) == JuMP.Model
+      @test JuMP.solver_name(C_rknn2_linear.tuning.modeler) == "SCIP"
+      @test length(JuMP.object_dictionary(C_rknn2_linear.tuning.modeler)) == 6
+      @test size(
+          JuMP.object_dictionary(C_rknn2_linear.tuning.modeler)[:u_reference],
+      ) == (QTP_sys_rknn2.inputdim, horizon)
+      @test size(JuMP.object_dictionary(C_rknn2_linear.tuning.modeler)[:u]) ==
+            (QTP_sys_rknn2.inputdim, horizon)
+      @test size(JuMP.object_dictionary(C_rknn2_linear.tuning.modeler)[:e_u]) ==
+            (QTP_sys_rknn2.inputdim, horizon)
+      @test size(
+          JuMP.object_dictionary(C_rknn2_linear.tuning.modeler)[:x_reference],
+      ) == (QTP_sys_rknn2.statedim, horizon + 1)
+      @test size(JuMP.object_dictionary(C_rknn2_linear.tuning.modeler)[:x]) ==
+            (QTP_sys_rknn2.statedim, horizon + 1)
+      @test size(JuMP.object_dictionary(C_rknn2_linear.tuning.modeler)[:e_x]) ==
+            (QTP_sys_rknn2.statedim, horizon + 1)
+      @test JuMP.objective_function(C_rknn2_linear.tuning.modeler) != 0 #to improve
+      @test JuMP.objective_function_type(C_rknn2_linear.tuning.modeler) ==
+            JuMP.QuadExpr
+      @test JuMP.list_of_constraint_types(C_rknn2_linear.tuning.modeler) == [
+          (AffExpr, MathOptInterface.EqualTo{Float64}),
+          (AffExpr, MathOptInterface.LessThan{Float64}),
+          (VariableRef, MathOptInterface.EqualTo{Float64}),
+      ]
+      @test C_rknn2_linear.tuning.reference == references
+      @test C_rknn2_linear.tuning.horizon == horizon
+      @test C_rknn2_linear.tuning.terminal_ingredient.Xf == "none"
+      @test C_rknn2_linear.tuning.terminal_ingredient.P != 0
+      @test C_rknn2_linear.tuning.sample_time == 5.0
+      @test C_rknn2_linear.tuning.max_time == 30 #to modify
+      @test size(C_rknn2_linear.initialization) ==
+            (QTP_sys_rknn2.statedim,)
+      @test size(C_rknn2_linear.computation_results.x) ==
+            (QTP_sys_rknn2.statedim, horizon + 1)
+      @test size(C_rknn2_linear.computation_results.e_x) ==
+            (QTP_sys_rknn2.statedim, horizon + 1)
+      @test size(C_rknn2_linear.computation_results.u) ==
+            (QTP_sys_rknn2.inputdim, horizon)
+      @test size(C_rknn2_linear.computation_results.e_u) ==
+            (QTP_sys_rknn2.inputdim, horizon)
+      ### end evaluate neuralnetODE_type1 L MPC implementation ###
+  
+  end
+
+@testset "design linear Model Predictive Control with rknn4 model" begin
+
+      hmin = 0.2
+      h1max = 1.36
+      h2max = 1.36
+      h3max = 1.30
+      h4max = 1.30
+      qmin = 0
+      qamax = 4
+      qbmax = 3.26
+  
+      #Constraint definition:
+      x_cons = LazySets.Hyperrectangle(
+          low = [hmin, hmin, hmin, hmin],
+          high = [h1max, h2max, h3max, h4max],
+      )
+      u_cons = LazySets.Hyperrectangle(low = [qmin, qmin], high = [qamax, qbmax])
+  
+      #get the resnet model to design the mpc controler
+      rknn4_machine =
+          machine("./models_saved/rknn4_train_result.jls")
+  
+      #extract best model from the all trained models
+      mlj_rknn4 =
+          fitted_params(fitted_params(rknn4_machine).machine).best_model
+      f_rknn4 =
+          fitted_params(fitted_params(rknn4_machine).machine).best_fitted_params[1]
+      type_rknn4 = mlj_rknn4.builder
+  
+      #system definition with MAthematical systems
+      QTP_sys_rknn4 =
+          MathematicalSystems.ConstrainedBlackBoxControlDiscreteSystem(
+              f_rknn4,
+              4,
+              2,
+              x_cons,
+              u_cons,
+          )
+  
+      #MPC design parameters
+      horizon = 15
+      sample_time = 5
+      terminal_ingredient = false
+      max_time = 5
+      x = [0.65, 0.65, 0.65, 0.65] * ones(1, horizon + 1)
+      u = [1.2, 1.2] * ones(1, horizon)
+      references = AutomationLabsModelPredictiveControl.ReferencesStateInput(x, u)
+  
+      ##############################
+      ###           neuralnetODE_type1          ###                    
+      ##############################
+  
+      ### neuralnetODE_type1 L MPC ###
+      method = AutomationLabsModelPredictiveControl.LinearProgramming()
+  
+      C_rknn4_linear = _model_predictive_control_design(
+          QTP_sys_rknn4,
+          horizon,
+          sample_time,
+          references;
+          mpc_programming_type = "linear",
+      )
+  
+      ### start evaluate neuralnetODE_type1 L MPC implementation ###
+      @test C_rknn4_linear.system == QTP_sys_rknn4
+      @test typeof(C_rknn4_linear.tuning.modeler) == JuMP.Model
+      @test JuMP.solver_name(C_rknn4_linear.tuning.modeler) == "SCIP"
+      @test length(JuMP.object_dictionary(C_rknn4_linear.tuning.modeler)) == 6
+      @test size(
+          JuMP.object_dictionary(C_rknn4_linear.tuning.modeler)[:u_reference],
+      ) == (QTP_sys_rknn4.inputdim, horizon)
+      @test size(JuMP.object_dictionary(C_rknn4_linear.tuning.modeler)[:u]) ==
+            (QTP_sys_rknn4.inputdim, horizon)
+      @test size(JuMP.object_dictionary(C_rknn4_linear.tuning.modeler)[:e_u]) ==
+            (QTP_sys_rknn4.inputdim, horizon)
+      @test size(
+          JuMP.object_dictionary(C_rknn4_linear.tuning.modeler)[:x_reference],
+      ) == (QTP_sys_rknn4.statedim, horizon + 1)
+      @test size(JuMP.object_dictionary(C_rknn4_linear.tuning.modeler)[:x]) ==
+            (QTP_sys_rknn4.statedim, horizon + 1)
+      @test size(JuMP.object_dictionary(C_rknn4_linear.tuning.modeler)[:e_x]) ==
+            (QTP_sys_rknn4.statedim, horizon + 1)
+      @test JuMP.objective_function(C_rknn4_linear.tuning.modeler) != 0 #to improve
+      @test JuMP.objective_function_type(C_rknn4_linear.tuning.modeler) ==
+            JuMP.QuadExpr
+      @test JuMP.list_of_constraint_types(C_rknn4_linear.tuning.modeler) == [
+          (AffExpr, MathOptInterface.EqualTo{Float64}),
+          (AffExpr, MathOptInterface.LessThan{Float64}),
+          (VariableRef, MathOptInterface.EqualTo{Float64}),
+      ]
+      @test C_rknn4_linear.tuning.reference == references
+      @test C_rknn4_linear.tuning.horizon == horizon
+      @test C_rknn4_linear.tuning.terminal_ingredient.Xf == "none"
+      @test C_rknn4_linear.tuning.terminal_ingredient.P != 0
+      @test C_rknn4_linear.tuning.sample_time == 5.0
+      @test C_rknn4_linear.tuning.max_time == 30 #to modify
+      @test size(C_rknn4_linear.initialization) ==
+            (QTP_sys_rknn4.statedim,)
+      @test size(C_rknn4_linear.computation_results.x) ==
+            (QTP_sys_rknn4.statedim, horizon + 1)
+      @test size(C_rknn4_linear.computation_results.e_x) ==
+            (QTP_sys_rknn4.statedim, horizon + 1)
+      @test size(C_rknn4_linear.computation_results.u) ==
+            (QTP_sys_rknn4.inputdim, horizon)
+      @test size(C_rknn4_linear.computation_results.e_u) ==
+            (QTP_sys_rknn4.inputdim, horizon)
+      ### end evaluate neuralnetODE_type1 L MPC implementation ###
+  
+  end
 end